Reorganize CuPy tests into multiple files
